create a teeiframe component
