+++ apps/nuxt/src/pages/iframe/projet/[projectSlug].vue
@@ -1,3 +1,12 @@
+<template>
+  <TeeIframe
+    :link="href"
+    :title="`Accédez aux aides publiques pour votre projet de ${projectTitle}`"
+    :image-src="currentProject?.image"
+    :image-alt="`image du projet ${projectTitle}`"
+  />
+</template>
+
 <script setup lang="ts">
 import { MiddlewareName } from '@/middleware/type/middlewareName'
 import Analytics from '@/tools/analytic/analytics'
@@ -9,11 +18,8 @@ definePageMeta({
   middleware: [MiddlewareName.hasProject]
 })
 
-const imageResizerChild = () => import('@iframe-resizer/child')
-imageResizerChild()
-
 const { currentProject } = storeToRefs(useProjectStore())
-const title = currentProject.value?.title?.toLowerCase() || ''
+const projectTitle = currentProject.value?.title?.toLowerCase() || ''
 const router = useRouter()
 const href = router.resolve({ name: RouteName.CatalogProjectDetail, params: { projectSlug: currentProject.value?.slug } }).href
 
@@ -22,98 +28,10 @@ if (import.meta.client) {
   const parentUrl = params.get('parent_url')
   Analytics.sendEvent('iframe_view', {
     type: 'project',
-    title: title,
+    title: projectTitle,
     referrer_url: parentUrl
   })
 }
 
 useHead(MetaRobots.noIndexFollow())
 </script>
-
-<template>
-  <div class="fr-container fr-px-0">
-    <div class="fr-grid-row">
-      <div class="fr-col-12">
-        <div class="fr-card fr-enlarge-link fr-card--horizontal fr-card--icon-link fr-card--grey fr-p-4v fr-card--no-border">
-          <div class="fr-card__body">
-            <div class="fr-card__content">
-              <h3 class="fr-card__title">
-                <a
-                  :href="href"
-                  target="_blank"
-                  rel="noopener external"
-                >
-                  Accédez aux aides publiques pour votre projet de {{ title }}
-                </a>
-              </h3>
-              <p class="fr-card__desc fr-text--md">
-                Le service public <strong>Transition écologique des entreprises</strong> vous permet de trouver les aides, accompagnements
-                et ressources issues de l'ensemble des acteurs publics pour vous aider à réaliser votre projet de transition écologique
-                (ADEME, CCI, CMA, Bpifrance, DGFiP, etc.)
-              </p>
-              <div class="fr-card__start"></div>
-            </div>
-          </div>
-          <div class="fr-card__header">
-            <div class="fr-card__img">
-              <div>
-                <img
-                  class="fr-responsive-img image-with-gradient"
-                  :src="currentProject?.image"
-                  :alt="`image / ${currentProject?.title}`"
-                />
-              </div>
-              <img
-                class="fr-card__logo"
-                src="/images/logos/mission-transition-logo-alone.png"
-                alt="Transition Ecologique des Entreprises - ADEME"
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</template>
-
-<style scoped lang="scss">
-.fr-card {
-  .fr-card__title {
-    a::after {
-      display: inline-block !important;
-    }
-  }
-
-  .fr-card__header {
-    flex-basis: 25%;
-  }
-
-  .fr-card__img {
-    img {
-      max-height: 250px;
-    }
-
-    div {
-      height: 100%;
-
-      &:first-child::after {
-        content: '';
-        position: absolute;
-        height: 100%;
-        width: 100%;
-        top: 0;
-        left: 0;
-        background: linear-gradient(to bottom right, rgb(255 255 255 / 0%) 70%, rgb(255 255 255 / 100%) 95%) !important;
-      }
-    }
-  }
-
-  .fr-card__logo {
-    position: absolute;
-    width: 50px;
-    bottom: 0;
-    height: 50px;
-    right: 0;
-  }
-}
-</style>
